+++ packages/nuxt-locokit/src/pages/auth/ResetPassword.vue
@@ -44,7 +44,7 @@ import { storeToRefs } from 'pinia'
 import { useI18n } from 'vue-i18n'
 import WithBackground from '../../layouts/WithBackground.vue'
 import { useStoreAuth } from '../../stores/auth'
-import { ROUTES_NAMES } from '../paths'
+import { ROUTES_NAMES } from '../../paths'
 import { ref, useHead, useRoute } from '#imports'
 
 const { t } = useI18n({ useScope: 'global' })
