+++ packages/nuxt-locokit/src/pages/auth/VerifySignup.vue
@@ -43,7 +43,7 @@ import { PasswordForm } from '@locokit/designsystem'
 import { storeToRefs } from 'pinia'
 import { useI18n } from 'vue-i18n'
 import WithBackground from '../../layouts/WithBackground.vue'
-import { ROUTES_NAMES } from '../paths'
+import { ROUTES_NAMES } from '../../paths'
 import { useStoreAuth } from '../../stores/auth'
 import { ref, useHead, useRoute } from '#imports'
